# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Title
  description: Title
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/mywebapp'
components:
  schemas:
    doctorObject:
      type: object
      properties:
        name:
          type: string
        speciality:
          type: string
    doctorsObjects:
      type: array
      items:
        $ref: "#/components/schemas/doctorObject"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string
paths:
  /doctors:
    get:
      summary: Метод получения списка докторов
      description: Get All Doctors
      tags:
        - Doctors
      operationId: getAllDoctors
      responses:
        '200':
          description: Список докторов
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/doctorsObjects"
        'default':
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Метод добавление доктора
      tags:
        - Doctors
      operationId: createDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/doctorObject"
      responses:
        '200':
          description: Доктор добавлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/doctorObject"
  /doctors/{id}:
    get:
      summary: Метод получения доктора по ID
      description: Get Doctor
      tags:
        - Doctors
      operationId: getDoctorById
      parameters:
        - name: doctor_id
          in: path
          required: true
          description: ID доктора
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/doctorObject"
    delete:
      summary: Метод удаления доктора по ID
      tags:
        - Doctors
      operationId: deleteDoctorById
      parameters:
        - name: doctor_id
          in: path
          required: true
          description: ID доктора
          schema:
            type: string
      responses:
        '200':
          description: Доктор удален
          content:
            application/json: {}
  /doctor/{id}:
    patch:
      summary: Метод обновления доктора
      description: Update Doctor
      tags:
        - Doctors
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: doctor
          in: query
          schema:
            $ref: "#/components/schemas/doctorObject"

      responses:
        '200':
          description: Doctor updated